database/requirements.txt
SQLAlchemy==2.0.32
python-dotenv==1.0.1